requirements.txt
daytona>=0.153.0
litellm>=1.35.0
gspread>=6.0.0
google-auth>=2.28.0
streamlit>=1.32.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
linkedin-api>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
